Youtube-Chatbot-main/Backend/requirements.txt
fastapi
uvicorn
youtube-transcript-api
langchain
langchain-openai
langchain-text-splitters
langchain-chroma
langchain-core
python-dotenv
translate
pydantic
chromadb